+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/comparison/CompareShrinkTrunkWithFixed.kt
@@ -1,4 +1,4 @@
-package org.cryptobiotic.rlauxe.integration
+package org.cryptobiotic.rlauxe.comparison
 
 import org.cryptobiotic.rlauxe.core.AlphaMart
 import org.cryptobiotic.rlauxe.core.FixedEstimFn
