+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/comparison/ComparisonChooseD.kt
@@ -1,10 +1,12 @@
-package org.cryptobiotic.rlauxe.sim
+package org.cryptobiotic.rlauxe.comparison
 
 import org.cryptobiotic.rlauxe.util.makeCvrsByExactMean
 import org.cryptobiotic.rlauxe.plots.geometricMean
 import org.cryptobiotic.rlauxe.rlaplots.SRT
 import org.cryptobiotic.rlauxe.rlaplots.SRTcsvReader
 import org.cryptobiotic.rlauxe.rlaplots.SRTcsvWriter
+import org.cryptobiotic.rlauxe.sim.AlphaComparisonTask
+import org.cryptobiotic.rlauxe.sim.RepeatedTaskRunner
 import org.junit.jupiter.api.Test
 import kotlin.collections.forEach
 import kotlin.math.max
