+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/corla/TestCorla.kt
@@ -133,7 +133,7 @@ class TestCorla {
         val reportedMeans = listOf(0.501, 0.502, 0.503, 0.504, 0.505, 0.506, 0.5075, 0.508, 0.51, 0.52, 0.53, 0.54, 0.55, 0.56, 0.58, 0.6,)
 
         val N = 10000
-        val ntrials = 10000
+        val ntrials = 1000
 
         val tasks = mutableListOf<CorlaTask>()
         var taskCount = 0
