+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/polling/Dvalues.kt
@@ -1,7 +1,9 @@
-package org.cryptobiotic.rlauxe.sim
+package org.cryptobiotic.rlauxe.polling
 
-import org.cryptobiotic.rlauxe.util.makeCvrsByExactMean
 import org.cryptobiotic.rlauxe.rlaplots.SRTcsvWriter
+import org.cryptobiotic.rlauxe.sim.PollingTask
+import org.cryptobiotic.rlauxe.sim.RepeatedTaskRunner
+import org.cryptobiotic.rlauxe.util.makeCvrsByExactMean
 import org.junit.jupiter.api.Test
 
 // explore values of d for both polling and comparison
@@ -58,7 +60,17 @@ class Dvalues {
                 val cvrs = makeCvrsByExactMean(N, theta)
                 cvrMeanDiffs.forEach { cvrMeanDiff ->
                     val cvrMean = theta - cvrMeanDiff
-                    tasks.add(PollingTask(taskIdx++, N, cvrMean, cvrMeanDiff, d = 0, cvrs = cvrs, useFixedEstimFn = true))
+                    tasks.add(
+                        PollingTask(
+                            taskIdx++,
+                            N,
+                            cvrMean,
+                            cvrMeanDiff,
+                            d = 0,
+                            cvrs = cvrs,
+                            useFixedEstimFn = true
+                        )
+                    )
                 }
             }
         }
@@ -73,40 +85,6 @@ class Dvalues {
         println("${results.size} results written to ${writer.filename}")
     }
 
-    // TODO setting the eta0Factor by hand
-    @Test
-    fun comparisonAlpha() {
-        val thetas = listOf(.505, .51, .52, .53, .54, .55, .575, .6, .65, .7)
-        // val thetas = listOf(.501, .502, .503, .504, .505, .506, .507, .508, .51, .52, .53, .54)
-        val cvrMeanDiffs = listOf(0.2, 0.1, 0.05, 0.025, 0.01, 0.005, 0.0, -.005, -.01, -.025, -.05, -0.1, -0.2)
-        val dlist = listOf(10, 50, 250, 1250)
-
-        val ns = listOf(5000, 10000, 20000)
-        val ntrials = 100
-
-        val tasks = mutableListOf<AlphaComparisonTask>()
-        var taskIdx = 0
-        dlist.forEach { d ->
-            thetas.forEach { theta ->
-                ns.forEach { N ->
-                    cvrMeanDiffs.forEach { cvrMeanDiff ->
-                        val cvrMean = theta - cvrMeanDiff
-                        val cvrs = makeCvrsByExactMean(N, cvrMean)
-                        tasks.add(AlphaComparisonTask(taskIdx++, N, cvrMean, cvrMeanDiff, eta0Factor=1.9, d = d, cvrs = cvrs))
-                    }
-                }
-            }
-        }
-
-        val writer = SRTcsvWriter("/home/stormy/temp/sim/dvalues/comparisonAlpha9.csv")
-
-        val runner = RepeatedTaskRunner()
-        val results =  runner.run(tasks, ntrials)
-
-        writer.writeCalculations(results)
-        writer.close()
-        println("${results.size} results written to ${writer.filename}")
-    }
 
 
 }
